pyproject.toml
[project]
name = "pg-workstation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]
